Allow to use vpr even if you are not in the root directory
Before executing git related commands, vpr will explicitly look for the `.git` directory in the current directory and each of it's parent, stopping at the earliest match. This way, similar to the git command line tools, vpr may be used in sub folders.

If vpr cannot find a `.git` directory in any of the parents, then it falls back to using the current directory.